moved identifier to dedicated module
